Class "Interpolation"

This class serves as a base for interpolation methods. Although it is not defined as virtual for technical reason, it should not be used as is and is only useful after having been extended.

Objects from the Class

Objects from the class should not be created

Slots

name:

Object of class "character" holding the name of the interpolation

color:

Object of class "character" holding the name or the code (\#RRGGBB) of the color of the interpolation on the plot

physicalPositions:

Object of class "vector" holding the valid physical position of the map on which the interpolation is calculated. This redundancy is necessary to cary out the plotting of the map.

rates:

Object of class "vector" holding the local recombination rate estimations at the location of each markers.

visible:

Object of class "logical", whether or not the interpolation should be plotted

persistent:

Object of class "logical", whether or not the interpolation should be saved when the map is saved to text file.

Methods

argList

signature(object = "Interpolation"): This function is used in the function createList (c.f. code examples).

color<-

signature(object = "Interpolation"): changes the color of the interpolation.

color

signature(object = "Interpolation"): returns the color of the interpolation.

interpolate

signature(object = "Interpolation", map = "MareyMap"): recalculate the interpolation.

name<-

signature(object = "Interpolation"): changes the name of the interpolation.

name

signature(object = "Interpolation"): returns the name of the interpolation.

persistent<-

signature(object = "Interpolation"): changes the persistence state of the interpolation.

persistent

signature(object = "Interpolation"): returns the persistence state of the interpolation.

plotModel

signature(object = "Interpolation"): this function plots only the model of the marey map calculated by the method.

plotRate

signature(object = "Interpolation"): this function plots the recombination rates calculated by the method.

query

signature(object = "Interpolation", pos = "integer"): returns the recombination rate calculated by this method on a given location on the sequence.

rates<-

signature(object = "Interpolation"): replaces the rates with a new vector.

rates

signature(object = "Interpolation"): returns the rates of the interpolation.

userParam

signature(object = "Interpolation"): function returning information about the parameters modifiable by the user (c.f. code example.)

visible<-

signature(object = "Interpolation"): changes the visibility of the interpolation.

visible

signature(object = "Interpolation"): returns the visibility of the interpolation

Author

Aurélie Siberchicot aurelie.siberchicot@univ-lyon1.fr and Clément Rezvoy

See also