Class "InterpolationParam"

Class holding informations about a parameter of an interpolation method. This class is used in the function userParam,Interpolation-method. These informations are used by the GUI to dynamically create the dialog to modify this parameter.

Objects from the Class

Objects can be created by calls of the form new("InterpolationParam", ...) or using the methodInterpolationParam().

Slots

paramName:

Object of class "character" holding the name of the parameter.

paramType:

Object of class "character" holding the type of the parameter, currently supported parameter types are: logical, numeric, integer and color.

paramDesc:

Object of class "character" giving a small description of the parameter's signification.

paramDefault:

Object of class "ANY", default value for the parameter.

paramValues:

Object of class "vector" containing the set of valid values for the parameter (can be NULL).

paramMin:

Object of class "ANY", a minimum value for the parameter, or minimal string length if the parameter is of type character

paramMax:

Object of class "ANY" , a minimum value for the parameter, or maximum string length if the parameter is of type character

paramFun:

Object of class "character", the nae of function that is to be called to update the parameter's value. ex if paramFun is foo the parameter's value will be set using foo(interpolation)<-new_value

Methods

paramDefault<-

signature(object = "InterpolationParam"): ...

paramDefault

signature(object = "InterpolationParam"): ...

paramDesc<-

signature(object = "InterpolationParam", value = "character"): ...

paramDesc

signature(object = "InterpolationParam"): ...

paramFun<-

signature(object = "InterpolationParam", value = "character"): ...

paramFun

signature(object = "InterpolationParam"): ...

paramMax<-

signature(object = "InterpolationParam"): ...

paramMax

signature(object = "InterpolationParam"): ...

paramMin<-

signature(object = "InterpolationParam"): ...

paramMin

signature(object = "InterpolationParam"): ...

paramName<-

signature(object = "InterpolationParam", value = "character"): ...

paramName

signature(object = "InterpolationParam"): ...

paramType<-

signature(object = "InterpolationParam", value = "character"): ...

paramType

signature(object = "InterpolationParam"): ...

paramValues<-

signature(object = "InterpolationParam", value = "vector"): ...

paramValues

signature(object = "InterpolationParam"): ...

Author

Aurélie Siberchicot aurelie.siberchicot@univ-lyon1.fr and Clément Rezvoy

See also

Examples

  nam <- InterpolationParam()
  paramName(nam) <- "name"
  paramDesc(nam) <- "The name of the interpolation.\n"
  paramType(nam) <- "character"
  paramDefault(nam) <- "default name"
  paramMin(nam) <- 1
  paramMax(nam) <- NULL  
  paramFun(nam) <- "name"

  vis <- InterpolationParam()
  paramName(vis) <- "visible"
  paramDesc(vis) <- "Whether the line is visible on the plot or not"
  paramType(vis) <- "logical"
  paramDefault(vis) <- TRUE
  paramFun(vis) <- "visible"
  
  sav <- InterpolationParam()
  paramName(sav) <- "persistent"
  paramDesc(sav) <- "Indicate if the interpolation is to be be kept \n\
  when the map is saved to text file"
  paramType(sav) <- "logical"
  paramDefault(sav) <- TRUE
  paramFun(sav) <- "persistent"
  
  col <- InterpolationParam()
  paramName(col) <- "line color"
  paramDesc(col) <- "color of the line"
  paramType(col) <- "color"
  paramDefault(col) <- "#000000"
  paramFun(col) <- "color"