Class "MMLoess". Interpolation method plug-in for MareyMap

This packages provides an interpolation method that estimates the recombination rates using loess. For more details on how this method works, see loess.

Objects from the Class

Objects can be created by calls of the form new("MMLoess", ...) or simply by calling MMLoess().

Slots

span:

Object of class "numeric", parameters that controls the degree of smoothing.

degree:

Object of class "integer", the degree of the polynomials to be used, up to 2.

model:

Object of class "ANY"

name:

Object of class "character", name of the interpolation.

color:

Object of class "character", color of the interpolation on the map.

physicalPositions:

Object of class "vector", physical positions of the markers of the map on which the interpolation is calculated.

rates:

Object of class "vector", the local estimation of the recombination rate at the position of each markers.

visible:

Object of class "logical", whether the interpolation should be plotted or not.

persistent:

Object of class "logical", whether the interpolation should be saved along with the map in the text file

Extends

Class "Interpolation", directly.

Methods

createOrder

signature(object = "MMLoess"): see Interpolation-class.

degree<-

signature(object = "MMLoess"): changes the degree of the polynomials.

degree

signature(object = "MMLoess"): returns the degree of the polynomials.

interpolate

signature(object = "MMLoess", map = "MareyMap"): see Interpolation-class.

plotModel

signature(object = "MMLoess"): see Interpolation-class.

plotRate

signature(object = "MMLoess"): see Interpolation-class.

query

signature(object = "MMLoess", pos = "numeric"): see Interpolation-class.

span<-

signature(object = "MMLoess"): changes the value of the span.

span

signature(object = "MMLoess"): returns the value of the span.

userParam

signature(object = "MMLoess"): see Interpolation-class.

Author

Aurélie Siberchicot aurelie.siberchicot@univ-lyon1.fr and Clément Rezvoy

See also

Examples

data(Homo_sapiens_male)
chr7 <- Homo_sapiens_male[["Chromosome 07"]]
itr1 <- MMLoess()
color(itr1) <- "red"
span(itr1) <- 0.05
chr7 <- chr7 + itr1