Class "MapCollection"

This class holds several MareyMap-class objects organised in sets (using MapSet-class object).

Objects from the Class

Objects can be created by calls of the form new("MapCollection", ...) or using the function MapCollection. You may as well use the default collection provided by the package (as data()).

Slots

sets:

Object of class "list" containing a MapSet-class for each set$* of the collection.

Methods

[

signature(x = "MapCollection"): ...

[[<-

signature(x = "MapCollection", i = "ANY", j = "ANY", value = "MapSet"): ...

[[<-

signature(x = "MapCollection", i = "ANY", j = "ANY", value = "MareyMap"): ...

[[

signature(x = "MapCollection"): obtain the MapSet corresponding to a set name

$

signature(x = "MapCollection"): subset by set name

+

signature(e1 = "MapCollection", e2 = "MareyMap"): adds a map to the collection

+

signature(e1 = "MapCollection", e2 = "MapSet"): adds a set t the collection

-

signature(e1 = "MapCollection", e2 = "character"): removes the map named e2 from the object e1.

coerce

signature(from = "MapCollection", to = "data.frame"): --> as(object,"data.frame")

setNames

signature(object = "MapCollection"): returns the list of the names of all the sets present in the collection

length

signature(object = "MapCollection"): returns the length of the collection

textFile

signature(object = "MapCollection", file = "character"): Writes the content of the collection to text file text.

Author

Aurélie Siberchicot aurelie.siberchicot@univ-lyon1.fr and Clément Rezvoy

See also

Examples

data(Homo_sapiens_mean)
map1 <- Homo_sapiens_mean[["Chromosome 01"]]
# do some useful changes to the map...
mapName(map1)<- "Chrom 1 (modified)"