Class "MapSet"

This class is a collection of MareyMap for a given set. Not that useful on its own, it is used for internal organization inside a MapCollection

Objects from the Class

Objects can be created by calls of the form new("MapSet", ...). Or using the function MapSet().

Slots

maps:

Object of class "list" containing all the MareyMap of the set.

setName:

Object of class "character" containing the name of the set

Methods

[

signature(x = "MapSet"): returns a map from the list

[[<-

signature(x = "MapSet", i = "ANY", j = "ANY", value = "MareyMap"): replaces a map in the list with a new one

[[

signature(x = "MapSet"): returns a map from the list.

$

signature(x = "MapSet"): access a map by name in the list.

+

signature(e1 = "MapSet", e2 = "MareyMap"): adds a map to the collection.

-

signature(e1 = "MapSet", e2 = "character"): removes the map named e2 from the object e1.

coerce

signature(from = "MapSet", to = "data.frame"): converts the specieMaps object into a data.frame.

mapNames

signature(object = "MapSet"): returns a vector containing the names of all the maps.

setName<-

signature(object = "MapSet", value = "character"): changes the name of the set. Also changes the named of all MareyMap objects inside the list.

setName

signature(object = "MapSet"): returns the name of the set.

length

signature(object = "MapCollection"): returns the length of the set

textFile

signature(object = "MapSet", file = "character"): Writes the content of the MapSet object to file.

Author

Aurélie Siberchicot aurelie.siberchicot@univ-lyon1.fr and Clément Rezvoy

See also

Examples

data(Homo_sapiens_mean)
set <- Homo_sapiens_mean
setName(set) <- "Human"
set[["Chromosome 01"]]
#> An object of class "MareyMap"
#> Slot "setName":
#> [1] "Homo_sapiens_mean"
#> 
#> Slot "mapName":
#> [1] "Chromosome 01"
#> 
#> Slot "markerNames":
#>   [1] "D1S2217"          "WIAF-768"         "WIAF-3885"       
#>   [4] "WIAF-984"         "D1S243"           "D1S468"          
#>   [7] "TSC0140173"       "TSC0903069"       "WIAF-3800"       
#>  [10] "D1S2845"          "TSC0609187"       "TSC0343078"      
#>  [13] "D1S2893"          "TSC0228807"       "TSC0278455"      
#>  [16] "D1S2660"          "D1S1608"          "D1S2132"         
#>  [19] "D1S3405"          "D1S2795"          "D1S2145"         
#>  [22] "D1S47"            "D1S2633"          "TSC0201877"      
#>  [25] "D1S2870"          "D1S253"           "D1S2731"         
#>  [28] "D1S2642"          "D1S214"           "D1S1646"         
#>  [31] "D1S2663"          "D1S2694"          "TSC0082273"      
#>  [34] "TSC0295775"       "D1S548"           "D1S2666"         
#>  [37] "TSC0118401"       "D1S508"           "D1S1612"         
#>  [40] "D1S1615"          "TSC0393286"       "D1S1160"         
#>  [43] "TSC1331517"       "TSC1221825"       "TSC1262026"      
#>  [46] "D1S450"           "D1S244"           "D1S2736"         
#>  [49] "D1S2213"          "D1S1635"          "D1S1151"         
#>  [52] "D1S2667"          "TSC1557441"       "TSC0209871"      
#>  [55] "TSC1557445"       "TSC0123411"       "D1S2740"         
#>  [58] "TSC1103976"       "TSC1426762"       "D1S489"          
#>  [61] "TSC0016890"       "D1S434"           "D1S2718"         
#>  [64] "TSC0036608"       "D1S1597"          "D1S228"          
#>  [67] "D1S402"           "D1S2834"          "D1S407"          
#>  [70] "D1S507"           "D1S2728"          "D1S3472"         
#>  [73] "D1S1176"          "D1S2672"          "D1S436"          
#>  [76] "TSC0127563"       "TSC1789370"       "TSC1101326"      
#>  [79] "D1S2697"          "D1S170"           "TSC0364190"      
#>  [82] "TSC0807158"       "D1S2144"          "D1S3669"         
#>  [85] "D1S1592"          "TSC0148330"       "D1S2826"         
#>  [88] "TSC0110324"       "D1S2644"          "TSC0613868"      
#>  [91] "TSC0613877"       "D1S3726"          "D1S483"          
#>  [94] "D1S552"           "D1S378"           "D1S2647"         
#>  [97] "D1S199"           "D1S2843"          "D1S2732"         
#> [100] "D1S3720"          "D1S478"           "TSC0841689"      
#> [103] "TSC0898414"       "TSC0461871"       "D1S2828"         
#> [106] "GGAA30B06"        "D1S2725"          "D1S2702"         
#> [109] "D1S1539"          "D1S2864"          "A134VB1"         
#> [112] "D1S2698"          "D1S458"           "D1S2734"         
#> [115] "D1S482"           "D1S2838"          "TSC0110217"      
#> [118] "D1S2620"          "TSC0204928"       "TSC0440794"      
#> [121] "WIAF-1577"        "D1S1177"          "D1S1152"         
#> [124] "D1S2674"          "D1S234"           "D1S1676"         
#> [127] "WIAF-4559"        "D1S2885"          "HMG17"           
#> [130] "GGAA2D04"         "D1S455"           "D1S2749"         
#> [133] "D1S2639"          "FGR"              "D1S2787"         
#> [136] "TSC0110133"       "D1S2884"          "D1S2854"         
#> [139] "WIAF-1677"        "D1S470"           "D1S1622"         
#> [142] "TSC0169042"       "TSC0235290"       "D1S493"          
#> [145] "D1S449"           "D1S247"           "D1S2781"         
#> [148] "D1S513"           "D1S1161"          "WIAF-3437"       
#> [151] "TSC0912567"       "D1S233"           "TSC0097025"      
#> [154] "D1S2765"          "D1S2832"          "TSC0982626"      
#> [157] "TSC0059559"       "TSC0110174"       "D1S2676"         
#> [160] "D1S241"           "D1S201"           "D1S2830"         
#> [163] "D1S2783"          "D1S2613"          "D1S393"          
#> [166] "D1S195"           "D1S2730"          "D1S3463"         
#> [169] "WIAF-944"         "D1S496"           "GATA137F01"      
#> [172] "D1S441"           "D1S2656"          "D1S1190"         
#> [175] "D1S2729"          "D1S516"           "D1S472"          
#> [178] "ATA4E01"          "D1S2637"          "D1S2723"         
#> [181] "TSC0746118"       "D1S255"           "TSC0458980"      
#> [184] "TSC0183511"       "D1S186"           "D1S380"          
#> [187] "D1S3755"          "D1S1157"          "D1S432"          
#> [190] "D1S1591"          "D1S2657"          "D1S2892"         
#> [193] "D1S1598"          "D1S2131"          "TSC1231554"      
#> [196] "D1S2743"          "TSC0243108"       "D1S2632"         
#> [199] "D1S2706"          "D1S1188"          "D1S391"          
#> [202] "WIAF-1504"        "WIAF-3095"        "D1S2722"         
#> [205] "D1S3721"          "D1S2130"          "D1S62"           
#> [208] "D1S2645"          "D1S463"           "D1S1586"         
#> [211] "D1S193"           "D1S2861"          "D1S443"          
#> [214] "WIAF-3825"        "D1S447"           "D1S211"          
#> [217] "D1S2733"          "D1S2713"          "GATA145F08"      
#> [220] "WIAF-4561"        "WIAF-1934"        "D1S421"          
#> [223] "D1S2802"          "D1S451"           "D1S2677"         
#> [226] "D1S3175"          "WIAF-2162"        "D1S2797"         
#> [229] "D1S3714"          "WIAF-774"         "D1S2720"         
#> [232] "D1S2874"          "D1S2879"          "D1S3315"         
#> [235] "D1S2134"          "TSC0192338"       "TSC1223394"      
#> [238] "D1S2824"          "TSC0393352"       "D1S2724"         
#> [241] "TSC0999697"       "D1S2748"          "D1S197"          
#> [244] "D1S427"           "D1S1661"          "D1S232"          
#> [247] "D1S319"           "D1S231"           "D1S386"          
#> [250] "D1S1173"          "D1S509"           "D1S2661"         
#> [253] "TSC0024751"       "D1S417"           "TSC1727565"      
#> [256] "TSC0160740"       "TSC0078081"       "D1S2652"         
#> [259] "D1S475"           "D1S200"           "D1S2742"         
#> [262] "D1S2690"          "WIAF-640"         "TSC0502081"      
#> [265] "TSC0006496"       "D1S2867"          "TSC0405592"      
#> [268] "D1S2665"          "D1S2890"          "D1S1150"         
#> [271] "TSC0806097"       "TSC0244184"       "D1S476"          
#> [274] "TSC0806106"       "D1S2650"          "D1S2869"         
#> [277] "D1S405"           "D1S2648"          "D1S2752"         
#> [280] "D1S2700"          "D1S220"           "WIAF-1768"       
#> [283] "D1S2831"          "D1S2741"          "D1S3734"         
#> [286] "D1S2801"          "D1S2770"          "D1S2873"         
#> [289] "D1S3728"          "TSC0530115"       "D1S203"          
#> [292] "TSC0886398"       "TSC0009745"       "D1S2737"         
#> [295] "TSC0609480"       "D1S2822"          "D1S2846"         
#> [298] "D1S2788"          "D1S473"           "D1S209"          
#> [301] "WIAF-3293"        "TSC0312159"       "D1S246"          
#> [304] "TSC0396733"       "D1S230"           "TSC0289448"      
#> [307] "D1S2835"          "D1S2638"          "TSC1072442"      
#> [310] "D1S438"           "D1S515"           "D1S3467"         
#> [313] "D1S2617"          "D1S2684"          "D1S2710"         
#> [316] "D1S3724"          "D1S2754"          "D1S2825"         
#> [319] "D1S2886"          "D1S2866"          "D1S198"          
#> [322] "WIAF-198"         "WIAF-846"         "TSC1210558"      
#> [325] "D1S2806"          "D1S410"           "D1S2178"         
#> [328] "D1S368"           "TSC0766654"       "WIAF-1417"       
#> [331] "WIAF-3565"        "D1S2803"          "D1S448"          
#> [334] "D1S1162"          "WIAF-3233"        "D1S219"          
#> [337] "D1S411"           "D1S1603"          "D1S2798"         
#> [340] "D1S462"           "D1S2897"          "D1S501"          
#> [343] "D1S1178"          "D1S192"           "D1S2895"         
#> [346] "D1S224"           "D1S494"           "D1S1180"         
#> [349] "D1S1195"          "D1S1648"          "D1S3736"         
#> [352] "D1S1665"          "D1S2761"          "WIAF-2921"       
#> [355] "D1S464"           "D1S481"           "D1S2855"         
#> [358] "D1S532"           "TSC1218723"       "D1S216"          
#> [361] "D1S499"           "TSC0434999"       "TSC1096284"      
#> [364] "WIAF-1999"        "TSC0559391"       "D1S2614"         
#> [367] "D1S3733"          "D1S2876"          "D1S2841"         
#> [370] "D1S2618"          "D1S500"           "D1S465"          
#> [373] "D1S430"           "TSC0612838"       "TSC0597864"      
#> [376] "D1S1585"          "TSC0807161"       "TSC0807172"      
#> [379] "D1S2862"          "D1S2856"          "TSC1024552"      
#> [382] "TSC0600204"       "TSC0727574"       "D1S207"          
#> [385] "D1S454"           "D1S551"           "D1S1159"         
#> [388] "TSC0906070"       "TSC1267194"       "D1S2882"         
#> [391] "D1S226"           "D1S208"           "D1S2807"         
#> [394] "D1S488"           "D1S2774"          "D1S1541"         
#> [397] "D1S375"           "D1S2889"          "WIAF-3891"       
#> [400] "WIAF-1854"        "WIAF-866"         "D1S401"          
#> [403] "D1S2766"          "TSC0501132"       "WIAF-3042"       
#> [406] "D1S2865"          "TSC0825772"       "TSC0825777"      
#> [409] "TSC0608039"       "D1S2627"          "WIAF-1748"       
#> [412] "D1S435"           "TSC0757956"       "TSC1130274"      
#> [415] "TSC0080169"       "TSC1316843"       "WIAF-4050"       
#> [418] "TSC0720441"       "D1S1588"          "TSC0393393"      
#> [421] "D1S188"           "D1S1170"          "UTR_03754_X73110"
#> [424] "D1S424"           "D1S2804"          "D1S406"          
#> [427] "D1S2776"          "D1S2868"          "D1S2849"         
#> [430] "D1S2779"          "WIAF-3232"        "D1S236"          
#> [433] "WIAF-4566"        "WIAF-1842"        "D1S2775"         
#> [436] "D1S2813"          "WIAF-1642"        "D1S2819"         
#> [439] "D1S2664"          "D1S2719"          "D1S2793"         
#> [442] "D1S1163"          "D1S497"           "D1S415"          
#> [445] "D1S420"           "D1S2753"          "D1S1587"         
#> [448] "GATA124C08"       "TSC0093908"       "D1S2739"         
#> [451] "D1S2808"          "D1S1629"          "D1S540"          
#> [454] "TSC1226703"       "D1S1658"          "TSC0905377"      
#> [457] "TSC0372237"       "D1S2767"          "D1S2671"         
#> [460] "D1S1154"          "D1S223"           "D1S206"          
#> [463] "D1S1174"          "D1S2896"          "D1S486"          
#> [466] "D1S495"           "D1S2626"          "D1S535"          
#> [469] "TSC0988450"       "D1S2699"          "TSC0955700"      
#> [472] "TSC1081170"       "D1S2888"          "D1S485"          
#> [475] "D1S1631"          "D1S429"           "D1S2759"         
#> [478] "GATA133A08"       "D1S239"           "D1S2688"         
#> [481] "D1S1627"          "ATA42G12"         "D1S248"          
#> [484] "D1S3723"          "D1S1169"          "D1S2778"         
#> [487] "WIAF-2007"        "D1S2792"          "SS730565"        
#> [490] "D1S2651"          "TSC0129461"       "D1S221"          
#> [493] "TSC0906757"       "WIAF-3798"        "D1S2695"         
#> [496] "D1S457"           "D1S2708"          "D1S2726"         
#> [499] "D1S2809"          "TSC0912668"       "D1S2789"         
#> [502] "TSC0612515"       "WIAF-1310"        "D1S2837"         
#> [505] "D1S187"           "TSC0341936"       "TSC0167604"      
#> [508] "TSC0256134"       "D1S502"           "D1S418"          
#> [511] "D1S2746"          "D1S487"           "WIAF-2150"       
#> [514] "D1S2756"          "D1S3461"          "WIAF-2676"       
#> [517] "D1S2881"          "D1S1675"          "D1S1191"         
#> [520] "WIAF-3038"        "D1S250"           "D1S2852"         
#> [523] "D1S467"           "D1S2687"          "D1S189"          
#> [526] "D1S2744"          "D1S252"           "TSC0441948"      
#> [529] "D1S440"           "TSC0618308"       "TSC1320306"      
#> [532] "D1S2820"          "D1S2784"          "D1S453"          
#> [535] "TSC0220539"       "D1S2669"          "D1S2875"         
#> [538] "D1S2863"          "D1S534"           "D1S185"          
#> [541] "D1S514"           "D1S2696"          "D1S2344"         
#> [544] "D1S442"           "TSC0390477"       "TSC0390476"      
#> [547] "D1S1156"          "WIAF-3272"        "D1S1557"         
#> [550] "D1S2612"          "D1S2222"          "D1S3466"         
#> [553] "D1S498"           "D1S2347"          "D1S2345"         
#> [556] "D1S2343"          "TSC1002302"       "WIAF-1956"       
#> [559] "D1S2400"          "WIAF-858"         "D1S2346"         
#> [562] "D1S2715"          "D1S2858"          "TSC0719371"      
#> [565] "D1S305"           "D1S2714"          "D1S1153"         
#> [568] "D1S2777"          "D1S303"           "D1S1595"         
#> [571] "D1S2721"          "WIAF-2182"        "D1S2624"         
#> [574] "D1S1179"          "D1S506"           "D1S394"          
#> [577] "D1S1653"          "TSC0927437"       "SPTA1"           
#> [580] "WIAF-3580"        "WIAF-3583"        "D1S1187"         
#> [583] "WIAF-3921"        "TSC0013978"       "TSC0208809"      
#> [586] "TSC0395457"       "D1S2635"          "D1S398"          
#> [589] "CRP"              "D1S1167"          "D1S2707"         
#> [592] "TSC0598177"       "D1S2771"          "TSC0300223"      
#> [595] "TSC0813741"       "D1S484"           "TSC0606425"      
#> [598] "D1S2705"          "WIAF-4066"        "WIAF-4529"       
#> [601] "APOA2"            "D1S2675"          "D1S2768"         
#> [604] "TSC0917454"       "TSC0029764"       "D1S2844"         
#> [607] "TSC0345808"       "D1S104"           "D1S403"          
#> [610] "D1S400"           "D1S2628"          "D1S426"          
#> [613] "TSC0757140"       "D1S2878"          "D1S194"          
#> [616] "D1S382"           "ATA38A05"         "D1S2681"         
#> [619] "D1S2673"          "TSC1074394"       "D1S1158"         
#> [622] "TSC0265196"       "D1S2762"          "D1S2630"         
#> [625] "D1S61"            "D1S397"           "D1S318"          
#> [628] "D1S196"           "TSC0997356"       "D1S2750"         
#> [631] "TSC1076000"       "D1S431"           "D1S445"          
#> [634] "D1S2799"          "TSC0119083"       "TSC0594455"      
#> [637] "TSC1789609"       "D1S2658"          "D1S433"          
#> [640] "WIAF-3365"        "WIAF-2818"        "D1S3464"         
#> [643] "TSC0276483"       "D1S3719"          "WIAF-1121"       
#> [646] "WIAF-3460"        "WIAF-2644"        "D1S2851"         
#> [649] "D1S452"           "D1S210"           "D1S2815"         
#> [652] "D1S1619"          "TSC0191361"       "TSC0066556"      
#> [655] "D1S1165"          "TSC0056911"       "TSC0049924"      
#> [658] "D1S2790"          "TSC0002090"       "WIAF-4543"       
#> [661] "WIAF-4237"        "D1S2814"          "WIAF-1018"       
#> [664] "D1S1589"          "D1S242"           "D1S218"          
#> [667] "D1S2691"          "D1S2394"          "D1S416"          
#> [670] "D1S370"           "D1S2643"          "D1S2887"         
#> [673] "D1S2769"          "D1S2659"          "WIAF-3263"       
#> [676] "D1S2786"          "D1S480"           "D1S1640"         
#> [679] "TSC0980921"       "TSC0601404"       "D1S2791"         
#> [682] "TSC0611110"       "D1S212"           "D1S215"          
#> [685] "D1S2751"          "D1S2883"          "D1S2640"         
#> [688] "WIAF-2441"        "D1S2619"          "D1S466"          
#> [691] "TSC0809661"       "D1S2818"          "WIAF-2122"       
#> [694] "D1S376"           "D1S2623"          "D1S158"          
#> [697] "D1S2701"          "TSC1071379"       "TSC0059748"      
#> [700] "TSC1101340"       "D1S2127"          "D1S240"          
#> [703] "D1S2711"          "D1S444"           "D1S254"          
#> [706] "D1S191"           "D1S2848"          "D1S1194"         
#> [709] "D1S2138"          "CPLA3610"         "D1S3460"         
#> [712] "D1S202"           "D1S1642"          "D1S518"          
#> [715] "D1S1604"          "GATA134C08"       "D1S222"          
#> [718] "TSC0804564"       "D1S238"           "TSC0669135"      
#> [721] "D1S3470"          "TSC0962530"       "TSC0962526"      
#> [724] "D1S2877"          "D1S461"           "D1S428"          
#> [727] "D1S492"           "D1S1189"          "D1S2823"         
#> [730] "D1S384"           "D1S422"           "D1S3731"         
#> [733] "D1S3468"          "D1S2625"          "D1S408"          
#> [736] "D1S533"           "D1S412"           "TSC0662919"      
#> [739] "D1S1614"          "TSC0953663"       "TSC0990306"      
#> [742] "D1S2757"          "TSC0601498"       "D1S2794"         
#> [745] "GATA135F02"       "D1S1183"          "D1S2816"         
#> [748] "TSC0253271"       "TSC0253273"       "TSC1006288"      
#> [751] "D1S2840"          "D1S1660"          "D1S413"          
#> [754] "D1S2745"          "D1S1726"          "D1S3732"         
#> [757] "D1S1175"          "D1S1632"          "D1S2853"         
#> [760] "TSC1512046"       "D1S2817"          "TSC1512058"      
#> [763] "D1S2622"          "TSC0142117"       "TSC0790692"      
#> [766] "TSC1259826"       "D1S1181"          "TSC0361090"      
#> [769] "WIAF-270"         "WIAF-2636"        "D1S2716"         
#> [772] "D1S2738"          "D1S477"           "WIAF-3247"       
#> [775] "D1S1723"          "D1S306"           "D1S2764"         
#> [778] "D1S1171"          "D1S2615"          "D1S1606"         
#> [781] "D1S1727"          "D1S1647"          "D1S2655"         
#> [784] "D1S1724"          "D1S2686"          "D1S2683"         
#> [787] "D1S2760"          "WIAF-2078"        "D1S1678"         
#> [790] "D1S510"           "D1S1725"          "D1S2717"         
#> [793] "WIAF-3214"        "TSC0361528"       "D1S2668"         
#> [796] "TSC0970723"       "TSC0892277"       "TSC0379509"      
#> [799] "D1S1620"          "D1S504"           "D1S2872"         
#> [802] "D1S2773"          "WIAF-3122"        "D1S456"          
#> [805] "WIAF-1470"        "D1S249"           "D1S2156"         
#> [808] "D1S2636"          "TSC1174545"       "TSC0992291"      
#> [811] "D1S2772"          "WIAF-3762"        "D1S2735"         
#> [814] "WIAF-952"         "D1S2727"          "D1S3465"         
#> [817] "D1S2796"          "D1S2782"          "WIAF-2149"       
#> [820] "D1S2685"          "TSC0302500"       "D1S2692"         
#> [823] "WIAF-883"         "TSC0440960"       "TSC0996741"      
#> [826] "WIAF-2847"        "D1S2891"          "TSC0490553"      
#> [829] "D1S245"           "TSC1053651"       "TSC0111105"      
#> [832] "D1S471"           "D1S491"           "TSC0000177"      
#> [835] "D1S70"            "WIAF-234"         "D1S205"          
#> [838] "D1S2812"          "D1S414"           "D1S2810"         
#> [841] "WIAF-1779"        "D1S2780"          "D1S425"          
#> [844] "D1S505"           "D1S1667"          "D1S217"          
#> [847] "D1S2703"          "D1S2646"          "TSC0973555"      
#> [850] "TSC0226253"       "D1S419"           "D1S237"          
#> [853] "D1S2857"          "TSC0912535"       "TSC0681871"      
#> [856] "D1S2141"          "TSC0326231"       "D1S2629"         
#> [859] "D1S474"           "D1S2827"          "D1S229"          
#> [862] "TSC0521024"       "TSC0450542"       "D1S490"          
#> [865] "TSC0370258"       "D1S227"           "D1S395"          
#> [868] "D1S2860"          "D1S1610"          "D1S399"          
#> [871] "D1S2621"          "D1S2616"          "D1S549"          
#> [874] "D1S2758"          "D1S2880"          "D1S1185"         
#> [877] "D1S2641"          "D1S1626"          "D1S2689"         
#> [880] "D1S2894"          "D1S320"           "D1S2871"         
#> [883] "TSC0912135"       "D1S2859"          "TSC0912136"      
#> [886] "D1S2821"          "D1S469"           "D1S2763"         
#> [889] "D1S439"           "D1S479"           "D1S3722"         
#> [892] "D1S1186"          "TSC0297426"       "D1S1644"         
#> [895] "TSC0845872"       "D1S389"           "D1S1601"         
#> [898] "WIAF-834"         "D1S1617"          "D1S2847"         
#> [901] "D1S2631"          "D1S2805"          "D1S2833"         
#> [904] "D1S103"           "D1S1656"          "D1S225"          
#> [907] "D1S251"           "D1S3462"          "WIAF-1013"       
#> [910] "WIAF-1014"        "D1S2709"          "TSC0275013"      
#> [913] "WIAF-2014"        "D1S178"           "D1S437"          
#> [916] "TSC0125650"       "TSC0834358"       "TSC0613375"      
#> [919] "D1S459"           "WIAF-815"         "TSC0035703"      
#> [922] "TSC0193267"       "D1S2800"          "TSC0409706"      
#> [925] "TSC0241298"       "TSC0132719"       "D1S446"          
#> [928] "D1S1540"          "D1S163"           "D1S2712"         
#> [931] "D1S2649"          "D1S235"           "WIAF-4029"       
#> [934] "D1S8"             "TSC0273348"       "TSC1200125"      
#> [937] "TSC0297521"       "D1S2680"          "D1S2850"         
#> [940] "TSC0498713"       "D1S2678"          "WIAF-2072"       
#> [943] "WIAF-200"         "TSC0762837"       "D1S517"          
#> [946] "TSC0836400"       "D1S2670"          "TSC1034091"      
#> [949] "D1S1149"          "TSC0454916"       "TSC1105002"      
#> [952] "D1S1594"          "D1S2785"          "D1S304"          
#> [955] "D1S180"           "TSC0928181"       "TSC0839829"      
#> [958] "TSC0839843"       "D1S184"           "TSC0839839"      
#> [961] "D1S204"           "WIAF-537"         "D1S547"          
#> [964] "D1S321"           "TSC0443128"       "D1S2842"         
#> [967] "TSC0198897"       "D1S2679"          "D1S404"          
#> [970] "D1S1590"          "D1S2811"          "TSC0964965"      
#> [973] "D1S409"           "D1S3410"          "TSC0275619"      
#> [976] "TSC1255407"       "D1S102"           "D1S2693"         
#> [979] "D1S423"           "TSC0277904"       "D1S2836"         
#> [982] "SS2642891"        "TSC1035001"       "TSC1582780"      
#> [985] "TSC1086099"       "D1S2215"          "D1S2682"         
#> [988] "D1S3739"         
#> 
#> Slot "physicalPositions":
#>   [1]    837022    919830    989675   2128800   2171436   3608114   3632202
#>   [8]   3669448   3769273   4368931   4406051   4451168   4502024   4512461
#>  [15]   4523792   4715023   4914098   5046353   5414487   5510834   5690193
#>  [22]   5996052   6011645   6212680   6224030   6293893   6635605   6784663
#>  [29]   6896284   7076712   7191860   7276031   7308093   7336746   7377105
#>  [36]   7389054   7402880   7541650   8052251   8503578   8551104   8598740
#>  [43]   9298900   9305177   9335548   9519686  10508330  10549930  10779276
#>  [50]  10921820  11398827  11421304  11422487  11423944  11428839  11472243
#>  [57]  11855266  11887792  11980329  11982317  12004558  12266684  12563423
#>  [64]  12802717  13529493  13731716  13910182  14209796  14599435  14774076
#>  [71]  14905643  14946247  14955083  15024000  15615821  16115620  16119962
#>  [78]  16158597  16164847  17042848  17444643  17497757  17528973  17528974
#>  [85]  17812843  18158062  18178730  18725705  18771806  18774852  18793238
#>  [92]  18887766  18977465  19012272  19259794  19573442  19702324  20255868
#>  [99]  20381654  20657234  21345033  21364853  21391837  21423906  21687777
#> [106]  21725243  21875126  22293740  22465934  22618334  22656775  23014137
#> [113]  23182354  23585312  23598574  23606444  23607959  23608846  23624265
#> [120]  23632241  23946106  23971391  24234452  24595247  24896719  25146208
#> [127]  25579384  25823032  26480896  26663784  26973508  26979870  27621563
#> [134]  27639112  27811118  29011006  29032074  29098373  29339337  29827858
#> [141]  29980047  29980217  29988982  30101860  30388854  30544775  30622774
#> [148]  31000823  31080714  31082012  31133420  31153069  31169752  31736577
#> [155]  31970710  32847263  33002197  33127896  33154807  33168955  33692215
#> [162]  33842325  34024426  34330439  34622262  34788585  34886897  34952750
#> [169]  34985143  35076566  35103940  35346463  35863545  36348509  36740144
#> [176]  36796899  37102836  37143699  37176153  37176656  37291934  37318809
#> [183]  37377432  37384360  37900315  38163539  38621373  38864285  39046253
#> [190]  39215111  39714051  39860012  40011254  40317270  40341501  40383999
#> [197]  40391888  40609706  40741387  41013741  41049173  41162204  41167114
#> [204]  41222563  41395585  41486604  41665388  41925671  42181838  42446383
#> [211]  42688385  42775954  43257561  43359532  43626647  43900448  44165862
#> [218]  44182569  44487069  44786393  44789058  44925914  45265076  45426418
#> [225]  45461176  45686724  45763120  46645612  47394953  47406100  47619903
#> [232]  47620251  47626326  47645209  47993288  48394415  48429828  48474741
#> [239]  48922590  48947197  48962059  50074010  50462518  50957579  51142730
#> [246]  51430369  51617860  51915220  52433525  52831843  53520181  54046737
#> [253]  54993986  55025814  55099658  55164918  55172757  55178866  55596513
#> [260]  55722183  56405454  56867560  57107041  57272952  57283503  57320190
#> [267]  57363518  57405154  57585421  57872895  57937893  57946780  57950992
#> [274]  57957961  58249001  58472528  58520831  58611189  58826156  58829066
#> [281]  58998327  59041042  59290966  59437912  59442378  59613401  59670609
#> [288]  60165900  60383883  60965068  60993914  60994110  61036834  61038086
#> [295]  61121085  61244709  61311901  61569001  61618844  61684243  62229943
#> [302]  62233326  62272163  62291693  62314305  62338392  63039179  63264990
#> [309]  63509848  63528694  64054298  64117936  64185011  64285716  64375536
#> [316]  64853626  64941578  65038286  65220009  65549828  66723188  66872905
#> [323]  66928718  67567068  67580177  67851141  67886951  67943565  67949300
#> [330]  68303274  68542722  68630296  68883469  69158693  69461257  69553067
#> [337]  69766778  69820433  70376846  70382329  70432121  70616518  71158903
#> [344]  71198209  71340040  71852241  72915312  72931467  73118928  73125690
#> [351]  73896392  73941412  74828481  74914733  75443971  75591804  75793020
#> [358]  77204458  77347169  77367345  77501948  77529992  77726063  77737572
#> [365]  77801420  78320021  78631510  78696174  79199315  79386510  79681696
#> [372]  80886503  80989624  81110460  81696179  81710618  81726083  81763852
#> [379]  81925286  82069640  82074216  82076450  82102754  82255460  82480150
#> [386]  82607317  82627705  82669866  82671872  82749888  82806251  82875029
#> [393]  82980047  82997682  83172207  83245686  83294892  84547865  84824201
#> [400]  85353456  85629584  85681145  85784306  85804414  86614828  88063114
#> [407]  88092488  88122298  88142058  88622275  89898487  91270791  91281751
#> [414]  91310987  91316615  91836500  91860257  91902487  91926272  91953752
#> [421]  92207881  92254080  92305077  92592219  92593695  92739051  92922089
#> [428]  93047614  93231833  93412488  94498535  94522428  94642366  94707840
#> [435]  94957295  94975540  95106373  95237609  95657916  96530101  96805710
#> [442]  96943063  97703255  98190118  98209272  98230642  98241588  98397368
#> [449]  98531558  98543626  98969350  99379010  99428848  99549578  99552248
#> [456]  99557075  99572306  99679805 100978446 101092479 101309426 101397358
#> [463] 101529504 101684654 101859814 102273381 102707656 104088561 104483584
#> [470] 104492351 104543674 104565239 104801722 104979869 105372673 105404910
#> [477] 105556834 105993533 106551465 106651059 106675707 106787669 106867024
#> [484] 107366613 108759955 108759960 109192708 109452305 109850578 109911032
#> [491] 110004898 110013765 110036453 110293822 110408149 110590892 110774286
#> [498] 110896307 111126402 111704865 111711950 111754278 111754425 111987181
#> [505] 112072510 112163176 112199876 112219606 112220047 112680598 112737827
#> [512] 112872308 112926139 112967681 113338541 114410136 114427372 114452195
#> [519] 114477667 114604000 114923163 115066207 115709327 115914121 116405887
#> [526] 116983930 117268814 117269273 117427978 117515864 117546954 117553726
#> [533] 117554106 117876905 118012792 118214747 118476074 118789415 119390306
#> [540] 119887937 119980798 120212465 143020990 143119549 143191159 143212783
#> [547] 143432231 144344089 144447145 144900677 146700480 146997276 148114637
#> [554] 148134858 148346339 148346969 148373226 148485731 149818172 149925879
#> [561] 149982605 150397430 150857504 151052356 151095016 151928592 152081776
#> [568] 152273437 152450744 152501463 152966521 153376971 153444389 153637933
#> [575] 153658531 154132490 154745847 154818104 155393932 155405528 155405589
#> [582] 155707429 155802000 155902788 155959381 155961586 155983315 156451982
#> [589] 156497051 156822137 156885524 157394580 157408615 157517038 157576443
#> [596] 157580640 157626816 157671027 157781282 158002009 158005846 158941651
#> [603] 159244305 159548241 159625974 159680484 159688624 160368565 160458941
#> [610] 161208464 161880386 162038699 162068858 162135135 162168929 162413743
#> [617] 162580414 162785003 163312857 163321151 163327702 163347935 163687795
#> [624] 163947795 164135720 164241226 164269958 164335786 164419779 164461606
#> [631] 164477617 164551532 164686058 164824338 164825737 164839701 164849524
#> [638] 164986136 165099835 165359202 165552406 165780290 165825758 166345637
#> [645] 166392439 166559516 166622232 167048513 167255305 167592439 168446588
#> [652] 168468962 168495179 168531993 168533237 168579499 169745077 169755765
#> [659] 169800081 169852240 170608218 170704081 170868031 170992759 171103911
#> [666] 171234791 171730322 171859358 171925776 172248302 172254516 173072497
#> [673] 173361028 173403516 173554034 174019894 174078704 174451480 174488161
#> [680] 174560393 174583365 174629632 174814255 176360219 176742816 176836833
#> [687] 177458831 178505675 179007877 179035256 179069331 179070627 179342766
#> [694] 179460031 179605216 180872678 181354289 181385198 181397250 181440275
#> [701] 181470117 181921612 182249540 182487904 182511362 182553279 182566278
#> [708] 182720976 183097483 183519153 183632128 183697017 183962885 184281945
#> [715] 184366784 184366849 184502054 184840282 184877855 185566479 185725318
#> [722] 185776736 185778857 186076309 186150078 186196896 186291959 186849174
#> [729] 187227187 187322644 188110500 188256804 188272501 189000484 190671866
#> [736] 190789589 190797599 190809281 190843947 191390198 191430215 191469397
#> [743] 191543629 191988279 192319699 192880487 193382202 194700043 194724966
#> [750] 194787150 194989849 195343004 195352055 195863460 196007113 196030474
#> [757] 196263173 196263782 196383790 196784418 196808829 196817731 196846229
#> [764] 196850096 196874938 196935498 196985490 197000870 197179052 197835734
#> [771] 197852828 197878521 197985850 198011497 198122545 198136627 198317964
#> [778] 198649100 198649771 198731477 198915735 199289453 199296927 199319125
#> [785] 199368199 199803359 200020401 200187878 200263382 200327014 200700777
#> [792] 200792326 200803014 200805809 200813137 200846503 200850017 200860511
#> [799] 201295919 201376907 201557082 201749643 201767112 202190096 202257455
#> [806] 202447135 202501050 203023552 203092355 203128160 203370025 203404520
#> [813] 203480286 203580331 203625570 203747779 204147816 204185211 204357092
#> [820] 204378270 204423534 204427271 204437472 204454381 204466633 204579018
#> [827] 204840931 205946877 205994817 206023811 206041341 206105637 206131588
#> [834] 206152485 206169732 206176909 206499241 207316546 207431133 207793132
#> [841] 208387263 208455726 208471585 209006164 209159029 209436629 209477860
#> [848] 210446616 210761352 210792276 210926460 211117719 211141334 211534806
#> [855] 211561772 211583695 211627299 211834066 212065554 212527604 213482098
#> [862] 213505981 213520910 213539030 213547438 213682834 213694983 213868459
#> [869] 213915586 213915586 214736641 215588988 216032045 216344882 216925635
#> [876] 216953717 217522804 217677267 218019407 218187415 218328070 218544754
#> [883] 218715205 218718908 218745732 219096370 219330103 220768408 222084356
#> [890] 222516414 222640246 222693948 222807566 222819306 222914627 223931712
#> [897] 224535356 225907071 226010876 226172229 226503603 226999827 227054969
#> [904] 227143599 227212040 227466952 228021032 228257251 228309067 228309127
#> [911] 228337174 228338550 228465414 228666827 228733288 228748390 228762138
#> [918] 228798772 228801391 229426686 230724358 230757894 230765113 230768286
#> [925] 230966864 230985451 230987452 231091906 231166107 231203227 231704873
#> [932] 232219807 232466778 232586063 232883812 232895912 232927776 232955688
#> [939] 232956711 232961831 232967666 233037627 233040957 235631277 235686999
#> [946] 235717811 235875218 236410617 236473891 236479944 236509843 237106041
#> [953] 237202811 237614681 237691263 237857552 237867929 237883430 237890928
#> [960] 237903135 237932332 237970753 238081978 238147568 239176697 239198759
#> [967] 239215533 239316541 239361232 239423585 240017167 240373365 240391776
#> [974] 240394699 240426443 240488735 240492491 240626384 241871791 241928116
#> [981] 243196150 243389250 243851004 243873499 243904482 243928230 244456343
#> [988] 245358417
#> 
#> Slot "geneticDistances":
#>   [1]   0.00   0.00   0.00   5.80   5.80  11.03  11.06  12.51  12.51  13.46
#>  [11]  13.46  13.79  13.79  13.79  13.79  14.75  15.45  15.45  15.67  15.67
#>  [21]  15.67  17.13  17.13  18.31  18.59  19.17  19.17  19.17  19.17  19.49
#>  [31]  20.04  20.41  20.41  20.41  20.58  20.58  20.58  21.55  21.70  23.13
#>  [41]  23.54  24.48  24.64  24.97  24.97  24.97  25.01  25.11  26.75  26.75
#>  [51]  27.77  27.77  27.77  27.77  27.77  27.77  29.51  29.51  29.83  29.83
#>  [61]  29.83  29.83  31.99  31.99  32.28  32.92  33.79  34.16  35.29  35.29
#>  [71]  35.69  35.69  35.96  36.08  38.08  38.08  38.08  38.32  38.67  39.87
#>  [81]  39.87  41.20  41.20  41.20  42.02  42.63  43.06  45.33  45.33  45.33
#>  [91]  45.33  46.40  46.71  46.87  46.97  47.05  47.05  48.01  48.50  48.50
#> [101]  49.33  49.33  49.33  49.33  50.14  50.22  50.28  50.66  50.66  51.09
#> [111]  51.09  52.88  52.88  52.88  52.88  52.88  52.88  52.88  52.88  52.88
#> [121]  52.88  54.37  54.60  54.60  54.72  55.10  55.10  55.10  55.72  56.09
#> [131]  56.62  56.62  57.05  57.05  57.34  58.01  58.01  58.01  58.86  58.86
#> [141]  59.02  59.02  59.02  59.15  59.36  60.04  60.04  61.50  62.23  62.23
#> [151]  62.23  62.23  62.23  62.23  62.88  62.88  62.88  63.04  63.04  63.04
#> [161]  63.89  64.61  64.61  65.97  66.00  66.00  66.34  66.34  66.34  66.34
#> [171]  66.34  66.34  66.34  66.55  67.80  68.07  68.26  68.37  68.37  68.37
#> [181]  68.72  68.72  70.03  70.03  70.05  70.83  72.65  72.65  72.65  72.65
#> [191]  73.58  73.58  73.58  74.02  74.48  74.48  74.48  75.20  75.20  75.20
#> [201]  76.12  76.12  76.12  76.12  76.12  76.12  76.16  76.72  76.72  76.72
#> [211]  76.90  76.90  78.10  78.10  78.10  78.10  78.10  78.10  78.10  78.10
#> [221]  78.10  79.35  79.35  79.35  79.35  79.35  79.35  79.35  79.72  80.17
#> [231]  80.58  80.58  80.58  80.88  81.37  81.65  81.65  81.65  81.89  81.95
#> [241]  81.97  81.97  82.15  82.28  82.28  82.28  82.28  82.61  82.61  82.80
#> [251]  83.81  84.78  86.51  86.51  86.51  86.64  86.64  86.64  87.39  87.39
#> [261]  87.64  88.49  89.58  89.58  89.58  90.13  90.13  90.30  90.30  90.68
#> [271]  91.16  91.16  91.16  91.16  91.65  91.76  91.76  92.29  92.43  92.49
#> [281]  92.49  92.49  93.03  93.78  93.78  93.99  94.35  94.75  94.75  95.20
#> [291]  95.20  95.20  95.20  95.20  95.64  95.89  95.89  96.77  96.77  97.07
#> [301]  97.89  97.89  97.89  97.89  97.89  97.89  98.40  98.85  99.13  99.13
#> [311] 100.28 100.28 100.28 100.28 100.87 101.16 101.31 101.31 101.31 101.31
#> [321] 102.69 103.38 103.38 104.29 104.29 104.47 104.69 104.69 104.69 104.69
#> [331] 104.69 105.57 106.09 106.09 106.09 106.29 106.29 106.82 106.82 106.82
#> [341] 106.82 106.82 107.05 107.47 107.53 108.05 108.06 108.21 108.30 108.30
#> [351] 108.30 108.30 109.13 109.32 109.32 109.44 109.44 111.30 111.30 111.30
#> [361] 111.30 111.66 111.66 111.66 111.66 111.94 112.04 112.04 112.64 113.10
#> [371] 113.10 113.84 114.08 114.08 114.72 114.72 114.89 116.19 116.19 116.28
#> [381] 116.28 116.28 116.28 116.28 116.28 116.48 116.48 116.61 116.61 116.61
#> [391] 116.61 116.77 116.97 116.97 116.97 116.97 116.97 117.61 117.79 117.98
#> [401] 119.06 119.06 119.37 119.37 119.37 121.44 122.32 122.32 122.32 122.32
#> [411] 122.32 125.27 125.27 125.27 125.27 125.27 125.43 125.69 126.11 126.11
#> [421] 126.44 126.44 126.44 126.44 126.44 126.44 126.44 126.63 126.63 126.63
#> [431] 128.27 128.30 128.30 128.87 128.87 128.87 128.87 128.87 129.13 129.66
#> [441] 129.87 129.87 130.03 130.42 130.42 130.42 130.42 130.42 131.20 131.20
#> [451] 131.56 131.56 131.90 132.04 132.04 132.04 132.04 132.61 133.78 133.78
#> [461] 133.78 133.78 133.78 133.93 133.93 135.56 135.56 136.19 136.40 136.40
#> [471] 136.40 136.40 136.40 137.03 137.07 137.42 137.42 137.76 138.27 138.27
#> [481] 138.27 138.27 138.27 139.23 139.70 140.16 140.25 140.90 140.98 140.98
#> [491] 140.98 140.98 140.98 140.98 142.44 142.44 143.02 143.55 143.55 144.13
#> [501] 144.13 144.13 144.13 144.69 144.71 145.04 145.04 145.04 145.04 145.76
#> [511] 145.76 145.76 146.24 146.29 146.29 147.22 147.22 147.22 147.22 147.97
#> [521] 148.03 148.03 148.29 149.05 149.54 150.61 150.78 150.78 151.29 151.29
#> [531] 151.29 151.29 151.29 151.69 151.69 151.69 151.69 152.50 152.50 152.50
#> [541] 152.83 153.40 153.98 153.98 154.29 154.29 154.29 154.29 154.29 154.29
#> [551] 155.24 156.54 156.54 156.54 157.06 157.06 157.06 157.48 158.66 158.66
#> [561] 158.66 158.93 159.08 159.20 159.20 160.41 160.41 160.41 160.41 160.41
#> [571] 160.41 160.41 161.76 161.76 161.76 161.76 163.77 164.38 164.38 164.38
#> [581] 164.74 164.74 165.40 165.40 165.74 165.74 165.74 165.94 165.94 167.82
#> [591] 167.98 167.98 168.46 169.57 169.57 169.57 169.67 169.67 170.25 170.25
#> [601] 170.74 170.89 172.02 172.61 172.92 172.92 172.92 174.33 174.33 174.33
#> [611] 177.42 177.42 177.42 177.42 177.77 177.77 177.77 178.63 178.99 179.34
#> [621] 179.34 179.34 179.34 179.34 180.00 181.38 181.38 181.38 181.65 181.65
#> [631] 181.74 181.74 181.74 181.88 181.88 181.88 181.88 182.42 182.42 182.46
#> [641] 183.70 183.70 183.70 183.70 183.70 184.44 184.44 185.17 185.17 185.82
#> [651] 186.50 186.50 186.50 186.50 186.50 186.50 187.10 187.13 187.13 187.13
#> [661] 188.09 188.09 188.09 188.09 188.09 188.09 188.18 188.63 188.63 188.66
#> [671] 189.05 189.65 189.91 189.91 190.28 190.28 190.28 190.57 190.80 190.80
#> [681] 190.80 190.80 191.32 191.83 191.91 192.11 192.11 194.96 194.96 195.15
#> [691] 195.15 195.26 195.99 195.99 195.99 197.05 197.05 197.05 197.05 197.05
#> [701] 197.87 197.87 198.03 198.30 198.30 198.30 198.30 198.30 198.59 198.86
#> [711] 198.88 198.89 199.16 199.44 199.44 199.44 199.52 199.52 199.91 200.67
#> [721] 200.67 200.67 200.67 200.67 200.67 200.67 200.67 201.30 201.30 201.30
#> [731] 201.30 201.30 201.30 201.70 203.86 203.86 203.86 204.13 204.13 204.13
#> [741] 204.13 204.13 204.35 204.39 204.64 204.64 204.78 205.81 205.81 205.81
#> [751] 205.81 205.81 205.81 206.05 206.62 206.62 206.62 206.62 206.62 206.62
#> [761] 206.62 206.62 206.75 207.12 207.12 207.91 207.91 207.91 207.91 207.94
#> [771] 207.96 208.17 208.72 208.80 208.97 209.16 209.59 209.59 210.22 211.52
#> [781] 211.52 211.52 211.52 211.52 211.65 211.65 211.65 212.18 213.76 214.36
#> [791] 214.36 214.36 214.36 214.36 214.36 215.05 215.05 215.05 215.08 215.28
#> [801] 216.24 216.24 216.24 216.53 216.99 216.99 216.99 218.42 218.42 218.99
#> [811] 219.28 219.28 219.28 219.28 219.28 219.46 219.68 219.68 219.68 219.68
#> [821] 219.68 219.68 219.68 219.68 219.68 220.20 220.43 220.43 222.64 222.64
#> [831] 222.64 222.64 222.75 222.75 222.95 222.95 223.29 224.33 224.33 224.33
#> [841] 224.33 224.87 224.91 225.68 225.68 225.78 225.78 227.73 228.09 228.09
#> [851] 228.09 228.09 228.40 229.17 229.17 229.17 229.17 229.17 229.17 229.62
#> [861] 231.98 231.98 231.98 231.98 231.98 233.24 233.24 233.24 233.24 233.24
#> [871] 233.24 234.49 234.77 235.05 235.58 235.83 236.15 236.15 236.15 236.15
#> [881] 236.60 236.60 236.79 236.79 236.79 237.15 237.15 239.51 239.53 239.53
#> [891] 239.53 239.53 239.53 239.53 239.53 239.53 239.53 243.08 243.08 243.08
#> [901] 243.08 244.13 244.13 244.50 244.72 244.72 245.13 245.42 245.42 245.42
#> [911] 245.42 246.58 246.58 246.58 246.58 246.58 247.04 247.04 247.04 247.58
#> [921] 250.83 250.83 250.83 250.83 251.78 251.95 251.95 252.08 252.17 252.17
#> [931] 253.98 253.98 253.98 254.96 254.96 254.96 255.86 256.38 256.40 256.40
#> [941] 256.40 258.55 258.55 263.64 263.64 263.64 263.89 265.45 265.45 265.45
#> [951] 265.45 267.20 268.13 269.03 269.03 269.30 269.30 269.30 269.30 269.30
#> [961] 269.68 269.68 269.68 269.88 273.08 273.08 273.30 273.30 273.30 274.33
#> [971] 274.33 274.33 274.33 274.33 274.33 274.33 274.33 274.33 279.89 279.89
#> [981] 282.87 282.87 283.78 283.78 283.78 283.94 284.94 286.53
#> 
#> Slot "markerValidity":
#>   [1] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [16] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [31] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [46] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [61] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [76] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#>  [91] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [106] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [121] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [136] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [151] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [166] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [181] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [196] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [211] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [226] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [241] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [256] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [271] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [286] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [301] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [316] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [331] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [346] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [361] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [376] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [391] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [406] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [421] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [436] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [451] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [466] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [481] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [496] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [511] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [526] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [541] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [556] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [571] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [586] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [601] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [616] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [631] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [646] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [661] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [676] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [691] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [706] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [721] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [736] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [751] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [766] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [781] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [796] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [811] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [826] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [841] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [856] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [871] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [886] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [901] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [916] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [931] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [946] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [961] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [976] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> 
#> Slot "interpolations":
#> list()
#>